[Refactor] UIComponent 모듈 분리 및 텍스트필드, 라벨, 버튼 코드 작성 및 위치 변경
