+++ HousUIComponent/Sources/HousUIComponent/Export.swift
@@ -2,8 +2,9 @@
 //  Export.swift
 //  
 //
-//  Created by 김지현 on 2023/05/22.
+//  Created by 김호세 on 2023/06/02.
 //
 
-@_exported import HousButton
-// write your new module below
+// MARK: - Globally import
+@_exported import ThirdPartyLibraryManager
+@_exported import SnapKit
