+++ HousUIComponent/Sources/HousUIComponent/Extension/Notification+.swift
@@ -1,13 +1,13 @@
 //
-//  Notification+Extension.swift
-//  Hous-iOS-release
+//  File.swift
+//  
 //
-//  Created by 김민재 on 2023/05/10.
+//  Created by 김호세 on 2023/06/02.
 //
 
 import Foundation
 
-extension NSNotification.Name {
+internal extension NSNotification.Name {
   static let change = Notification.Name("UITextFieldTextDidChangeNotification")
   static let beginEdit = Notification.Name("UITextFieldTextDidBeginEditingNotification")
   static let endEdit = Notification.Name("UITextFieldTextDidEndEditingNotification")
