+++ HousUIComponent/Sources/HousUIComponent/Label/HousLabel.swift
@@ -1,17 +1,10 @@
-//
-//  HousLabel.swift
-//  Hous-iOS-release
-//
-//  Created by 김민재 on 2023/05/10.
-//
-
 import UIKit
 
 import AssetKit
 
-final class HousLabel: UILabel {
+public final class HousLabel: UILabel {
 
-  init(text: String?, font: UIFont, textColor: UIColor) {
+  public init(text: String?, font: UIFont, textColor: UIColor) {
     super.init(frame: .zero)
     setStyle(text: text, font: font, textColor: textColor)
   }
